Match all certificates, not just the most specific
We are currently transitioning from ALBs that terminate the TLS to
NLBs move the termination through to Envoy. ALBs do not open the TLS
connection to the upstream using the same server name as was received
by that. This means the Envoy can't match on SNI and it fails out.

The documentation from envoy: https://www.envoyproxy.io/docs/envoy/latest/api-v2/api/v2/listener/listener.proto#listener-filterchainmatch
seems to indicate that the most specific FilterChaniMatch will rule
so we can differ this picking of which certificate to Envoy at runtime.